fix: relative URL in QR code is now absolute.
This fix uses the site header from the request to form an absolute URL.
Since this doesn't always work it would be best to require a hard coded
path somewhere. passing the request object all the way down to the
method rendering the QR code on the paper isn't really nice as well but
as long as there is now good other way and especially no setting
defining the server root this is the best way to go.